Fix server tck tests after latest 4.0.0 refactoring
