+++ test-suite-http-server-tck-azure-function-http/src/test/java/io/micronaut/http/server/tck/azurehttpfunction/AzureFunctionHttpServerUnderTest.java
@@ -17,8 +17,8 @@ import io.micronaut.http.HttpStatus;
 import io.micronaut.http.MediaType;
 import io.micronaut.http.MutableHttpResponse;
 import io.micronaut.http.client.exceptions.HttpClientResponseException;
-import io.micronaut.http.server.tck.ServerUnderTest;
 import io.micronaut.http.simple.SimpleHttpResponseFactory;
+import io.micronaut.http.tck.ServerUnderTest;
 
 import java.io.IOException;
 import java.io.UnsupportedEncodingException;
@@ -49,6 +49,11 @@ public class AzureFunctionHttpServerUnderTest implements ServerUnderTest {
         } catch (UnsupportedEncodingException e) {
             return new SimpleHttpResponseFactory().status(HttpStatus.INTERNAL_SERVER_ERROR);
         }
+    }
+
+    @Override
+    public <I, O, E> HttpResponse<O> exchange(HttpRequest<I> request, Argument<O> bodyType, Argument<E> errorType) {
+        return exchange(request, bodyType);
     }
 
     private <I> HttpRequestMessage<Optional<String>> adaptRequest(HttpRequest<I> request) throws UnsupportedEncodingException {
