+++ test-suite-http-server-tck-azure-function-http/src/test/resources/META-INF/services/io.micronaut.http.tck.ServerUnderTestProvider
